API References

Kitab.MetaData — Type

MetaData(author, ...)

Instantiate MetaData object

Kitab.OpenITIDB — Type

OpenITIDB(url::String)

Instantiate OpenITIDB type using url as input.

Kitab.OpenITIDB — Method

OpenITIDB(urls::Array{String})

Instantiate OpenITIDB type using multiple urls as input.

Base.delete! — Method

delete!(::Type{OpenITIDB}, row::Int64)

Delete a particular book file using input row as identifier from list of books that can be viewed using list(OpenITIDB)

Base.delete! — Method

delete!(::Type{OpenITIDB})

Delete the DB folder used for saving the downloaded OpenITIDB text books

Base.get — Method

get(::Type{OpenITIDB}, url::Array{String})

Download multiple OpenITI books using multiple urls through Array{String}.

Base.get — Method

get(::Type{OpenITIDB}, url::String)

Download the OpenITI data.

Kitab.extract — Method

extract(::Type{MetaData}, data::Array{String})

Extract metadata from raw data text.

Kitab.list — Method

list(::Type{OpenITIDB})

List all the downloaded OpenITI text books

Kitab.load — Method

load(::Type{OpenITIDB}, row::Int64)

Load the downloaded OpenITI data by specifying the row of the file in the output of list(OpenITIDB).