API References

Kitab.MetaData — Type

MetaData(author, ...)

Instantiate MetaData object

Kitab.OpenITIDB — Type

OpenITIDB(url::String)

Instantiate OpenITIDB type using url as input.

Kitab.OpenITIDB — Method

OpenITIDB(urls::Array{String})

Instantiate OpenITIDB type using multiple urls as input.

Base.delete! — Method

delete!(::Type{OpenITIDB})

Delete the db folder for saving the downloaded OpenITIDB text

Base.download — Method

download(openitis::Array{OpenITIDB})

Download OpenITIDB data using an array of it to downloaded multiple data.

Base.download — Method

download(openiti::OpenITIDB)

Download the OpenITI data.

Kitab.extract — Method

extract(::Type{MetaData}, data::Array{String})

Extract metadata from raw data text.

Kitab.list — Method

list(::Type{OpenITIDB})

List all the downloaded OpenITI data

Kitab.load — Method

load(::Type{OpenITIDB}, row::Int64)

Load the downloaded OpenITI data by specifying the row of the file in the output of list(OpenITIDB).